Fixed object linker extern issues
Previously there was an issue where the 'Extern' directive on the object linker would write into the stream directly.

The problem with this that if I want to add new external references, I need to increment the version of the file format.

To fix this each 'Extern' callsite specifies a unique ID.
If a callsite is added or removed, the linker is unaffected.