Compiler improvements
Callstack preamble is removed from compiled code when it is not needed.
Callstack table is allocated statically instead of on each function call.
'pushjmp()' code is inlined rather than needing to call a function.
Removed '__makeGUID' as it is no longer needed in emitted code.
Project module compiles required metatables in one place.
Replaced "jmp_0" with "popcall" for consistency.